requirements.txt
aniso8601==6.0.0
asn1crypto==0.24.0
astroid==2.2.5
cffi==1.12.3
Click==7.0
colorama==0.4.1
cryptography==2.6.1
Django==2.2.1
django-crontab==0.7.1
djangorestframework==3.9.4
Flask==1.0.3
Flask-JWT-Extended==3.18.2
Flask-RESTful==0.3.7
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.2
pycparser==2.19
Pygments==2.4.1
PyJWT==1.7.1
pylint==2.3.1
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
Werkzeug==0.15.4
wrapt==1.11.1
